задача на определение чётных чисел
